Alterado o tipo de dado da variável 'valor' de integer para double. O script do banco de dados também foi alterado para coincidir

+++ LeiloesTDSat/src/cadastroVIEW.java
@@ -145,7 +145,7 @@ public class cadastroVIEW extends javax.swing.JFrame {
         String valor = cadastroValor.getText();
         String status = "A Venda";
         produto.setNome(nome);
-        produto.setValor(Integer.parseInt(valor));
+        produto.setValor(Double.parseDouble(valor));
         produto.setStatus(status);
         ProdutosDAO produtodao = new ProdutosDAO();
         produtodao.cadastrarProduto(produto);
